updated, clean up, makefile optimizations

+++ src/scene/test.c
@@ -1,11 +1,13 @@
-#include <scene.h>
+#include <math.h>
 #include <utils.h>
+#include <MLX42/MLX42.h>
+#include <scene.h>
 
-double intersect_plane(t_objs *obj, const t_vec3 ray_dir, const t_vec3 ray_pos)
+float intersect_plane(t_objs *obj, const t_vec3 ray_dir, const t_vec3 ray_pos)
 {
     t_plane *plane = &obj->plane;
-    double intersection_distance;
-    double denom = dot_product(plane->direction, ray_dir); // Using dot product
+    float intersection_distance;
+    float denom = dot_product(plane->direction, ray_dir); // Using dot product
     if (fabs(denom) > 1e-6) {
         t_vec3 p0l0 = obj->coords - ray_pos; // Using vector subtraction
         intersection_distance = dot_product(p0l0, plane->direction) / denom; // Using dot product
@@ -16,20 +18,20 @@ double intersect_plane(t_objs *obj, const t_vec3 ray_dir, const t_vec3 ray_pos)
     return (-1);
 }
 
-double intersect_sphere(t_objs *obj, const t_vec3 ray_dir, const t_vec3 ray_pos)
+float intersect_sphere(t_objs *obj, const t_vec3 ray_dir, const t_vec3 ray_pos)
 {
-	double radius = obj->sphere.radius;
+	float radius = obj->sphere.radius;
 	t_vec3 oc = ray_pos - obj->coords; // Using vector subtraction
 
-	double a = dot_product(ray_dir, ray_dir); // Using dot product
-	double b = 2.0 * dot_product(oc, ray_dir); // Using dot product
-	double c = dot_product(oc, oc) - radius * radius; // Using dot product
-	double discriminant = b * b - 4 * a * c;
+	float a = dot_product(ray_dir, ray_dir); // Using dot product
+	float b = 2.0 * dot_product(oc, ray_dir); // Using dot product
+	float c = dot_product(oc, oc) - radius * radius; // Using dot product
+	float discriminant = b * b - 4 * a * c;
 
 	if (discriminant > 0)
 	{
-		double t0 = (-b - sqrt(discriminant)) / (2.0 * a);
-		double t1 = (-b + sqrt(discriminant)) / (2.0 * a);
+		float t0 = (-b - sqrt(discriminant)) / (2.0 * a);
+		float t1 = (-b + sqrt(discriminant)) / (2.0 * a);
 		if (t0 > 0 && t1 > 0)
 		{
 			return ((t0 < t1) ? t0 : t1); // Choose the closest positive t value
@@ -44,7 +46,7 @@ double intersect_sphere(t_objs *obj, const t_vec3 ray_dir, const t_vec3 ray_pos)
 
 void intersect_table(t_objs *obj, const t_vec3 ray_dir, const t_vec3 ray_pos)
 {
-	static double (*intersect_obj[NUM_OBJ_TYPES])(t_objs *, const t_vec3, const t_vec3) = {
+	static float (*intersect_obj[NUM_OBJ_TYPES])(t_objs *, const t_vec3, const t_vec3) = {
 		intersect_plane,
 		intersect_sphere
 	};
@@ -98,7 +100,7 @@ void render_scene(t_rt *rt, t_scene *scn)
 	const t_vec3 base_ray_dir = scn->camera.camera.ray_direction;
 	uint32_t bg_color = 0xFF0000;
 
-	double aspect_ratio = (double)WIDTH / HEIGHT;
+	float aspect_ratio = (float)WIDTH / HEIGHT;
 
 	for (int j = 0; j < HEIGHT; j++)
 	{
